Android - Gastón - Integration working.

+++ ButtonIntegration/android/app/src/main/java/com/buttonintegration/Expose/RNTImpressionView.java
@@ -2,8 +2,6 @@ package com.buttonintegration.Expose;
 
 import android.content.Context;
 import android.util.AttributeSet;
-import android.view.View;
-import android.widget.LinearLayout;
 
 import com.buttonintegration.Model.OfferDetail;
 import com.usebutton.sdk.impression.ImpressionView;
@@ -17,7 +15,7 @@ public class RNTImpressionView extends ImpressionView {
 
     // Button configuration function.
     public void configureWithDetails(OfferDetail detail) {
-        OfferDetails details = new OfferDetails.Builder(detail.getOfferDetailURL(), detail.getOfferDetailID(), detail.getOfferDetailRate(), detail.getOfferDetailType()).build();
+        OfferDetails details = new OfferDetails.Builder(detail.getOfferDetailURL(), detail.getOfferDetailID(), (float)detail.getOfferDetailRate(), detail.getOfferDetailType()).build();
         this.setCreativeType(detail.getOfferCreativeType());
         this.configureWith(details);
     }
